requirements.txt
pyyaml=6.0.1
termcolor=2.4.0
requests=2.31.0